requirements.txt
# MergeCraft Dependencies
# Python 3.11+ required

# Core ML
torch>=2.3.0
transformers>=4.45.0
deepspeed>=0.14.0
peft>=0.11.0
trl>=0.9.0
accelerate>=0.30.0
datasets>=2.20.0
bitsandbytes>=0.43.0
safetensors>=0.4.3

# vLLM Synthesis
vllm>=0.5.0
openai>=1.35.0

# GitHub data collection
requests>=2.31.0
httpx>=0.27.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
PyGithub>=2.3.0
ghapi>=1.0.4

# Code analysis
gitpython>=3.1.43
pygments>=2.18.0

# AI APIs
anthropic>=0.28.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=16.0.0
datasketch>=1.6.5

# Utilities
loguru>=0.7.0
tqdm>=4.66.0
rich>=13.7.0
typer>=0.12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pyyaml>=6.0.1
jsonlines>=4.0.0

# Serving
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
redis>=5.0.0

# Experiment tracking
wandb>=0.17.0
huggingface-hub>=0.24.0
